\name{snpPositions}
\alias{snpPositions-data}
\alias{snpPositions}
\docType{data}
\title{SNP Genomic Coordinates}
\description{
  Chromosome base positions of human SNPS, initially derived from the
  SNPS included on Illumina microarrays.
}
\usage{
data(snpPositions)
}
\format{A data frame (\code{snpPositions}), with two columns and 600,470
  rows. This object contains the postions of SNPs on the Illumina 600K
  chip. It can be used to simulate SNP-chip data.
}
\source{
  Produced by Mark Zucker from Illumina data.
}
\references{
  Zucker MR, Abruzzo LV, Herling CD, Barron LL, Keating MJ, Abrams ZB,
  Coombes KR. Inferring Clonal Heterogeneity in Cancer using SNP Arrays
  and Whole Genome Sequencing. Submitted.
}
\keyword{datasets}

