\name{hapmapSegments}
\alias{hapmapSegments-data}
\alias{hapmapSegments}
\docType{data}
\title{}
\description{
}
\usage{
data(hapmapSegments)
}
\format{A data frame (\code{hapmapSegments}), with seven columns and 61,163
  rows. This object contains the results of perfroming a sergmentaiton c
  copy number analysis (using \code{DNAcopy}) .

  Describe/interpret columns.
}
\source{
  Find the GEO source numbers for the raw data. Maybe in Carmen's paer.
}
\references{
  Schweighofer et al.

  Some HapMap reference.
}
\keyword{datasets}

