\name{snpPositions}
\alias{snpPositions-data}
\alias{snpPositions}
\docType{data}
\title{}
\description{
}
\usage{
data(snpPositions)
}
\format{A data frame (\code{snpPositions}), with two columns and 600,470
  rows. This object contains the postions of SNPs on the Illumina 600K
  chip. It can be used to simulate SNP-chip data.
}
\source{
}
\references{
}
\keyword{datasets}

